+++ apps/next13/app/auth/layout.tsx
@@ -1,0 +1,8 @@
+import {useAuthenticatedBlitzContext} from "@blitzjs/auth"
+
+export default async function RootLayout({children}: {children: React.ReactNode}) {
+  await useAuthenticatedBlitzContext({
+    redirectAuthenticatedTo: "/",
+  })
+  return <>{children}</>
+}