+++ apps/toolkit-app-passportjs/src/auth/mutations/resetPassword.ts
@@ -1,6 +1,7 @@
-import { SecurePassword, hash256 } from "@blitzjs/auth"
+import { hash256 } from "@blitzjs/auth"
 import db from "db"
 import { ResetPassword } from "../validations"
+import { SecurePassword } from "@blitzjs/auth/secure-password"
 import login from "./login"
 
 export class ResetPasswordError extends Error {