+++ packages/generator/src/prisma/field.ts
@@ -1,4 +1,5 @@
 import * as ast from "@mrleebo/prisma-ast"
+import {checkInputsOrRaise} from "../utils/checkInputOrRaise"
 import {capitalize, singlePascal, uncapitalize} from "../utils/inflector"
 
 export enum FieldType {
@@ -55,6 +56,8 @@ export class Field {
 
   // 'name:type?[]:attribute' => Field
   static parse(input: string, schema?: ast.Schema): Field[] {
+    checkInputsOrRaise(input)
+
     const [_fieldName, _fieldType = "String", _attribute] = input.split(":")
     let attribute = _attribute as string
     let fieldName = uncapitalize(_fieldName as string)
@@ -84,7 +87,7 @@ export class Field {
     }
     // use original unmodified field name in case the list handling code
     // has modified fieldName
-    if (isRelation(_fieldName as string)) {
+    if (typeof _fieldName === "string" && isRelation(_fieldName)) {
       // this field is an object type, not a scalar type
       const relationType = Relation[_fieldName]
       // translate the type into the name since they should stay in sync