+++ packages/generator/templates/app/test/index.test.tsx
@@ -4,19 +4,19 @@
 
  import { expect, vi, test } from "vitest"
  import { render } from "test/utils"
- 
+
  import Home from "../src/pages/index"
- 
+
  vi.mock("public/logo.png", () => ({
    default: {src: "/logo.png"}
  }))
- 
- 
+
+
  test.skip("renders blitz documentation link", () => {
    // This is an example of how to ensure a specific item is in the document
    // But it's disabled by default (by test.skip) so the test doesn't fail
-   // when you remove the the default content from the page
-   
+   // when you remove the default content from the page
+
    // This is an example on how to mock api hooks when testing
    vi.mock("src/users/hooks/useCurrentUser", () => (
     {
@@ -28,10 +28,9 @@
       })
     }
   ))
- 
+
    const { getByText } = render(<Home />)
-   const linkElement = getByText(/Documentation/i)
+   const linkElement = getByText(/Blitz Docs/i)
    expect(linkElement).toBeInTheDocument()
  })
- 
- 
+