+++ packages/blitz-next/CHANGELOG.md
@@ -1,5 +1,35 @@
 # @blitzjs/next
 
+## 2.0.0-beta.23
+
+### Patch Changes
+
+- Updated dependencies [c3c789740]
+  - blitz@2.0.0-beta.23
+  - @blitzjs/rpc@2.0.0-beta.23
+
+## 2.0.0-beta.22
+
+### Patch Changes
+
+- Updated dependencies [270361886]
+- Updated dependencies [989691ec8]
+- Updated dependencies [8aa22a0b2]
+  - @blitzjs/rpc@2.0.0-beta.22
+  - blitz@2.0.0-beta.22
+
+## 2.0.0-beta.21
+
+### Patch Changes
+
+- 10f98c681: Add an href property to the generated route manifest that will return a string of the pathname and included query params.
+- Updated dependencies [d692b4c1d]
+- Updated dependencies [0025856b9]
+- Updated dependencies [10f98c681]
+- Updated dependencies [d5b8faa86]
+  - @blitzjs/rpc@2.0.0-beta.21
+  - blitz@2.0.0-beta.21
+
 ## 2.0.0-beta.20
 
 ### Minor Changes